+++ changelog.md
@@ -10,6 +10,7 @@ This version is not yet released. If you are reading this on the website, then t
 - [`fill` `⬚`](https://uiua.org/docs/fill) can now be disabled for a function by filling with an empty list
 - [`parse` `⋕`](https://uiua.org/docs/parse) now has a glyph
 - [`sign` `±`](https://uiua.org/docs/sign), [`floor` `⌊`](https://uiua.org/docs/floor), [`ceiling` `⌈`](https://uiua.org/docs/ceiling), and [`round` `⁅`](https://uiua.org/docs/round) now work with [`under` `⍜`](https://uiua.org/docs/under)
+- [`reduce` `/`](https://uiua.org/docs/reduce) can now use a function that takes more than 2 arguments
 - Add some missing arithmetic inverses and unders involving [`flip` `:`](https://uiua.org/docs/flip)
 ### Interpreter
 - Make [`stack` `?`](https://uiua.org/docs/stack) and [`dump`](https://uiua.org/docs/dump) output show call stack